split rewriting in early and later query rewriting
