+++ go/vt/vtgate/planbuilder/early_rewriter.go
@@ -1,0 +1,62 @@
+/*
+Copyright 2021 The Vitess Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package planbuilder
+
+import (
+	"vitess.io/vitess/go/vt/sqlparser"
+	"vitess.io/vitess/go/vt/vtgate/semantics"
+)
+
+type earlyRewriter struct {
+	err      error
+	semTable *semantics.SemTable
+}
+
+func starRewrite(statement sqlparser.SelectStatement, semTable *semantics.SemTable) error {
+	r := earlyRewriter{
+		semTable: semTable,
+	}
+	sqlparser.Rewrite(statement, r.starRewrite, nil)
+	return r.err
+}
+
+func (r *earlyRewriter) starRewrite(cursor *sqlparser.Cursor) bool {
+	switch node := cursor.Node().(type) {
+	case *sqlparser.Select:
+		tables := r.semTable.GetSelectTables(node)
+		var selExprs sqlparser.SelectExprs
+		for _, selectExpr := range node.SelectExprs {
+			starExpr, isStarExpr := selectExpr.(*sqlparser.StarExpr)
+			if !isStarExpr {
+				selExprs = append(selExprs, selectExpr)
+				continue
+			}
+			starExpanded, colNames, err := expandTableColumns(tables, starExpr)
+			if err != nil {
+				r.err = err
+				return false
+			}
+			if !starExpanded {
+				selExprs = append(selExprs, selectExpr)
+				continue
+			}
+			selExprs = append(selExprs, colNames...)
+		}
+		node.SelectExprs = selExprs
+	}
+	return true
+}
